Ein bisschen eleganter
