+++ .gitignore
@@ -1,1 +1,1 @@
-*.pyc
+__pycache__/
