--- __init__.py
@@ -1,7 +1,0 @@
-# -*- coding: utf-8 -*-
-"""
-Created on Thu Jan 31 09:42:03 2019
-
-@author: NG7a8f3
-"""
-
