+++ meister_tools/__init__.py
@@ -1,0 +1,10 @@
+# -*- coding: utf-8 -*-
+"""
+Created on Thu Jan 31 09:42:03 2019
+
+@author: NG7a8f3
+"""
+
+from pathlib import Path
+
+path_db = Path(__file__).parent / 'odatastools.db'
