+++ meister_tools/__main__.py
@@ -1,0 +1,4 @@
+from .patrickstools2 import PTools
+
+app = PTools()
+app.mainloop()
